sep-19/requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
Automat==20.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
cssselect==1.1.0
Django==4.1.1
django-cors-headers==3.13.0
django-crontab==0.7.1
django-environ==0.9.0
djangorestframework==3.13.1
drf-yasg==1.21.3
fake-useragent==0.1.11
filelock==3.8.0
gevent==21.12.0
greenlet==1.1.3
grequests==0.6.0
h11==0.13.0
hyperlink==21.0.0
idna==3.4
incremental==21.3.0
inflection==0.5.1
itemadapter==0.7.0
itemloaders==1.0.6
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
lxml==4.9.1
MarkupSafe==2.1.1
outcome==1.2.0
packaging==21.3
parsel==1.6.0
Protego==0.2.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==0.21.0
pytz==2022.2.1
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
Scrapy==2.6.2
selenium==4.4.3
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tldextract==3.3.1
tqdm==4.64.1
trio==0.21.0
trio-websocket==0.9.2
Twisted==22.8.0
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
w3lib==2.0.1
webdriver-manager==3.8.3
wsproto==1.2.0
zope.event==4.5.0
zope.interface==5.4.0
